Refactoring - remove unnecessary class and renaming
